Extrapolating Parser from JSoupParser to improve extensibility.
