+++ mainmodule/src/main/java/com/sainsburys/integration/scraper/Handler.java
@@ -1,0 +1,7 @@
+package com.sainsburys.integration.scraper;
+
+public class Handler {
+	public void runScraper() {
+		
+	}
+}
