+++ mainmodule/src/main/java/com/sainsburys/integration/scraper/Main.java
@@ -3,7 +3,7 @@ package com.sainsburys.integration.scraper;
 public class Main {
 
 	public static void main(String[] args) {
-		System.out.println("Initial commit");
+		Handler handler = new Handler();
+		handler.runScraper();
 	}
-
 }
